Overhaul the integration tests (#66)
This simplifies the integration tests and allows us to more easily load
different seed data for different tests. The seed data is also simpler
now that we can do this, because we don't need a single database to
cater for all possible test cases.
