+++ data/seed/test/base/README.md
@@ -1,0 +1,4 @@
+
+# Base Test Seed Data
+
+This seed data includes a set up system with no authentication required. It's used to test the majority of the site without needing to worry about sessions and auth.
