+++ package.json
@@ -28,7 +28,7 @@
     "express": "^4.15.0",
     "express-session": "^1.15.2",
     "http-errors": "^1.6.1",
-    "knex": "^0.11.10",
+    "knex": "^0.13.0",
     "lodash": "^4.17.4",
     "morgan": "^1.8.1",
     "pg": "^4.5.6",
