+++ test/integration/api/get.js
@@ -1,11 +1,14 @@
-/* global agent */
+/* global agent, dashboard */
 'use strict';
+
+const loadSeedData = require('../helper/load-seed-data');
 
 describe('GET /api', () => {
 	let request;
 
 	beforeEach(() => {
 		request = agent.get('/api');
+		return loadSeedData(dashboard, 'base');
 	});
 
 	it('responds with a 302 status', done => {
