+++ test/integration/front-end/get-404.js
@@ -1,28 +1,14 @@
-/* global agent */
+/* global agent, dashboard */
 'use strict';
 
-describe('GET /', () => {
-	let request;
-
-	beforeEach(() => {
-		request = agent.get('/');
-	});
-
-	it('responds with a 200 status', done => {
-		request.expect(200).end(done);
-	});
-
-	it('responds with HTML', done => {
-		request.expect('Content-Type', 'text/html; charset=utf-8').end(done);
-	});
-
-});
+const loadSeedData = require('../helper/load-seed-data');
 
 describe('GET /404', () => {
 	let request;
 
 	beforeEach(() => {
 		request = agent.get('/404');
+		return loadSeedData(dashboard, 'base');
 	});
 
 	it('responds with a 404 status', done => {
